Detect code language in CSS style with pattern
When the language is specified as a CSS style with the pattern
`language-rust`, then it will be returned. This is applied when the
custom data attribute was not found first.

There are a few ways to specify the code language, either on the `<pre>`
block itself or when it contains a single nested `<code>` block (for
example: https://purplesyringa.moe/blog/the-expression-problem-and-rust/).